SeatacWeather {latticeExtra}	R Documentation

Daily Rainfall and Temperature at the Seattle-Tacoma Airport

Description

Daily Rainfall and Temperature at the Seattle-Tacoma Airport between January through March of 2007.

Usage

data(SeatacWeather)

Format

A data frame with 90 observations on the following 14 variables.

month

a factor with levels January, February, and March

day

day of the month

year

year, all 2007

max.temp

maximum temperature (Fahrenheit)

record.max

record maximum temperature

normal.max

normal maximum temperature

min.temp

minimum temperature

record.min

record minimum temperature

normal.min

normal minimum temperature

precip

precipitation (inches)

record.precip

record precipitation

normal.precip

normal precipitation

time.max

time of maximum temperature

time.min

time of minimum temperature

Details

The time of minimum and maximum temperatures should be interpreted as follows: the least two significant digits denote minutes (out of 60) and the next two significant digits denote hout (out of 24).

Source

http://www.atmos.washington.edu/cgi-bin/list_climate.cgi?clisea

[Package latticeExtra version 0.6-29 Index]