USAge {latticeExtra}	R Documentation

US national population estimates

Description

US national population estimates by age and sex from 1900 to 1979. The data is available both as a (3-dimensional) table and a data frame. The second form omits the 75+ age group to keep age numeric.

Usage

data(USAge.table)
data(USAge.df)

Format

USAge.table is a 3-dimensional array with dimensions

No 	 Name 	 Levels

1 	 Age 	 0, 1, 2, ..., 74, 75+

2 	 Sex 	 Male, Female

3 	 Year 	 1900, 1901, ..., 1979

Cells contain raw counts of estimated population.

USAge.df is a data frame with 12000 observations on the following 4 variables.

Age

a numeric vector, giving age in years

Sex

a factor with levels Male Female

Year

a numeric vector, giving year

Population

a numeric vector, giving population in millions

Details

The data for 1900-1929 are rounded to thousands. The data for 1900-1939 exclude the Armed Forces overseas and the population residing in Alaska and Hawaii. The data for 1940-1949 represent the resident population plus Armed Forces overseas, but exclude the population residing in Alaska and Hawaii. The data for 1950-1979 represent the resident population plus Armed Forces overseas, and also include the population residing in Alaska and Hawaii.

Source

U.S. Census Bureau website: http://www.census.gov/

The data were available as individual files for each year, with varying levels for the margins. The preprocessing steps used to reduce the data to the form given here are described in the scripts directory.

Examples

data(USAge.df)
head(USAge.df)

## Figure 10.7 from Sarkar (2008)
xyplot(Population ~ Age | factor(Year), USAge.df,
       groups = Sex, type = c("l", "g"),
       auto.key = list(points = FALSE, lines = TRUE, columns = 2),
       aspect = "xy", ylab = "Population (millions)",
       subset = Year %in% seq(1905, 1975, by = 10))

[Package latticeExtra version 0.6-29 Index]