USCancerRates {latticeExtra}	R Documentation

Rate of Death Due to Cancer in US Counties

Description

This data set records the annual rates of death (1999-2003) due to cancer by sex in US counties.

Usage

data(USCancerRates)

Format

A data frame with 3041 observations on the following 8 variables.

rate.male

a numeric vector, giving rate of death per 100,000 due to cancer among males

LCL95.male

a 95% lower confidence limit for rate.male

UCL95.male

a 95% upper confidence limit for rate.male

rate.female

a numeric vector, giving rate of death per 100,000 due to cancer among females

LCL95.female

a 95% lower confidence limit for rate.female

UCL95.female

a 95% upper confidence limit for rate.female

state

a factor with levels giving name of US state

county

a character vector giving county names, in a format similar to that used for county map boundaries in the maps package.

Details

See the scripts directory for details of data preprocessing steps.

From the website: Death data provided by the National Vital Statistics System public use data file. Death rates calculated by the National Cancer Institute using SEER*Stat. Death rates are age-adjusted to the 2000 US standard population [http://www.seer.cancer.gov/stdpopulations/stdpop.19ages.html]. Population counts for denominators are based on Census populations as modified by NCI.

Source

http://statecancerprofiles.cancer.gov/

Examples

data(USCancerRates)

[Package latticeExtra version 0.6-29 Index]