biocAccess {latticeExtra}	R Documentation

Hourly access attempts to Bioconductor website

Description

This data set records the hourly number of access attempts to the Bioconductor website (http://www.bioconductor.org) during January through May of 2007. The counts are essentially an aggregation of the number of entries in the access log.

Usage

data(biocAccess)

Format

A data frame with 3623 observations on the following 7 variables.

counts

the number of access attempts

day

the day of the month

month

a factor with levels Jan, Feb, ..., Dec

year

the year (all 2007)

hour

hour of the day, a numeric vector

weekday

a factor with levels Monday, Tuesday, ..., Sunday

time

a POSIXt representation of the start of the hour

Examples

data(biocAccess)
xyplot(stl(ts(biocAccess$counts[1:(24 * 30)], frequency = 24), "periodic"))

[Package latticeExtra version 0.6-29 Index]