ecdfplot {latticeExtra}	R Documentation

Trellis Displays of Empirical CDF

Description

Conditional displays of Empirical Cumulative Distribution Functions

Usage

ecdfplot(x, data, ...)

## S3 method for class 'formula'
ecdfplot(x, data,
         prepanel = "prepanel.ecdfplot", 
         panel = "panel.ecdfplot",
         ylab,
         ...)
## S3 method for class 'numeric'
ecdfplot(x, data = NULL, xlab, ...)

prepanel.ecdfplot(x, f.value = NULL, ...)

panel.ecdfplot(x, f.value = NULL, type = "s",
               groups = NULL, qtype = 7,
               ref = TRUE,
               ...)

Arguments

x
For ecdfplot, x is the object on which method dispatch is carried out. For the "formula" method, x is a formula describing the form of conditioning plot, and has to be of the form ~x, where x is assumed to be a numeric vector. Further conditioning variables are allowed as usual.

A similar interpretation holds for x in the "numeric" method as well as prepanel.ecdfplot and panel.ecdfplot.

data
For the "formula" method, a data frame containing values for any variables in the formula, as well as those in groups and subset if applicable.

prepanel, panel
panel and prepanel function used to create the display.

xlab, ylab
axis labels; typically a character string or an expression.

groups
a grouing variable of the same length as x. If specified, ECDF plots are computed for each subset defined by unique values of groups and the resulting functions superposed within each panel.

f.value, qtype
Defines how quantiles are calculated. See panel.qqmath.

ref
logical, whether a reference line should be drawn at 0 and 1

type
how the plot is rendered; see panel.xyplot

...
extra arguments, passed on as appropriate. Standard lattice arguments as well as arguments to panel.ecdfplot can be supplied directly in the high level ecdfplot call.

Value

ecdfplot produces an object of class "trellis". The update method can be used to update components of the object and the print method (usually called by default) will plot it on an appropriate plotting device.

Author(s)

Deepayan Sarkar deepayan.sarkar@r-project.org

See Also

qqmath for Quantile plots which are more generally useful, especially when comparing with a theoretical distribution other than uniform. An ECDF plot is essentially a transposed version (i.e., with axes switched) of a uniform quantile plot.

Examples

data(singer, package = "lattice")
ecdfplot(~height | voice.part, data = singer)

[Package latticeExtra version 0.6-29 Index]