gvhd10 {latticeExtra}	R Documentation

Flow cytometry data from five samples from a patient

Description

Flow cytometry data from blood samples taken from a Leukemia patient before and after allogenic bone marrow transplant. The data spans five visits.

Usage

data(gvhd10)

Format

A data frame with 113896 observations on the following 8 variables.

FSC.H

forward scatter height values

SSC.H

side scatter height values

FL1.H

intensity (height) in the FL1 channel

FL2.H

intensity (height) in the FL2 channel

FL3.H

intensity (height) in the FL3 channel

FL2.A

intensity (area) in the FL2 channel

FL4.H

intensity (height) in the FL4 channel

Days

a factor with levels -6 0 6 13 20 27 34

Source

http://web.archive.org/web/20070427103026/http://www.ficcs.org/software.html#Data_Files

References

Brinkman, R.R., et al. (2007). High-Content Flow Cytometry and Temporal Data Analysis for Defining a Cellular Signature of Graft-Versus-Host Disease. Biology of Blood and Marrow Transplantation 13–6

Examples

## Figure 3.4 from Sarkar (2008)
data(gvhd10)
histogram(~log2(FSC.H) | Days, gvhd10, xlab = "log Forward Scatter",
          type = "density", nint = 50, layout = c(2, 4))

[Package latticeExtra version 0.6-29 Index]