panel.key {latticeExtra}	R Documentation

Draw a simple key inside a panel of a lattice plot.

Description

Draw a simpleKey inside a panel of a lattice plot.

Usage

panel.key(text, ..., corner = c(0, 1), x = corner[1], y = corner[2])

Arguments

text, ...
entries in the simpleKey.

corner, x, y
defines the position of the key within the panel viewport. These are given in normalised coordinates between 0 and 1. The two elements of corner specify the x and y positions respectively.

See Also

simpleKey, draw.key, xyplot

Examples

xyplot(ozone ~ wind | equal.count(temperature, 2),
       data = environmental) +
   layer(panel.loess(x, y, span = 0.5), style = 1) +
   layer(panel.loess(x, y, span = 1.0), style = 2) +
   layer(panel.key(c("span = 0.5", "span = 1.0"), corner = c(1,.98),
                   lines = TRUE, points = FALSE), packets = 1)

[Package latticeExtra version 0.6-29 Index]