panel.qqmath.tails {latticeExtra}	R Documentation

Approximate distribution in qqmath but keep points on tails.

Description

Panel function for qqmath to reduce the number of points plotted by sampling along the specified distribution. The usual method for such sampling is to use the f.value argument to panel.qqmath. However, this panel function differs in two ways: (1) a specified number of data points are retained (not interpolated) on each tail of the distribution. (2) the sampling is evenly spaced along the specified distribution automatically (whereas f.value = ppoints(100) is evenly spaced along the uniform distribution only).

This function is deprecated as of lattice 0.18-4 (available for R 2.11.0). Use the tails.n argument of panel.qqmath instead.

Usage

panel.qqmath.tails(x, f.value = NULL, distribution = qnorm,
                    groups = NULL, ..., approx.n = 100, tails.n = 10)

Arguments

x, f.value, distribution, groups
see panel.qqmath.

...
further arguments passed on to panel.xyplot.

approx.n
number of points to use in approximating the distribution. Points will be equally spaced in the distribution space.

tails.n
number of points to retain (untouched) at both the high and low tails.

Author(s)

Felix Andrews felix@nfrac.org

See Also

panel.qqmath which should be used instead (as of lattice 0.18-4).

Examples

## see ?panel.qqmath

[Package latticeExtra version 0.6-29 Index]