xyplot.stl {latticeExtra}	R Documentation

Display stl fits with Lattice

Description

Display stl decomposition (seasonal, trend and irregular components using loess) with Lattice, like the base graphics function plot.stl.

Usage

## S3 method for class 'stl'
xyplot(x, data = NULL,
       outer = TRUE,
       layout = c(1, 4),
       strip = FALSE,
       strip.left = TRUE,
       as.table = TRUE,
       ylab = "",
       between = list(y = 0.5),
       panel =
       function(..., type) {
           if (packet.number() == 4) type <- "h"
           panel.xyplot(..., type = type)
       },
       ...)

Arguments

x
an stl object.

data
ignored.

outer, layout, strip, strip.left
passed to xyplot.ts.

as.table, ylab, between, panel, ...
passed to xyplot.ts.

Details

Unless strip.left is passed in explicitly, a custom strip will be drawn, where shaded bars are comparable across panels (representing the same data range).

Value

An object of class "trellis". The update method can be used to update components of the object and the print method (usually called by default) will plot it on an appropriate plotting device.

Author(s)

Deepayan Sarkar deepayan.sarkar@r-project.org

See Also

stl, xyplot.ts, xyplot

Examples

## example from ?stl
xyplot(stl(log(co2), s.window=21),
       main = "STL decomposition of CO2 data")

## alternative display
data(biocAccess)
xyplot(stl(ts(biocAccess$counts[1:(24 * 30)], frequency = 24), "periodic"),
       strip.left = "strip.default")
resizePanels()

## two different spans
xyplot(stl(nottem, s.window = 4)) +
as.layer(xyplot(stl(nottem, s.window = "periodic")), style = 2)

## components superposed
xyplot(stl(nottem, s.window = 4), superpose=TRUE,
  screens = list(data = "trend", trend = "trend", "residuals"),
  strip.left = TRUE, layout = c(1,2))

[Package latticeExtra version 0.6-29 Index]