ancestry {latticeExtra}	R Documentation

Modal ancestry by County according to US 2000 Census

Description

This data set records the population and the three most frequently reported ancestries by US county, according to the 2000 census.

Usage

data(ancestry)

Format

A data frame with 3219 observations on the following 5 variables.

county

A factor. An attempt has been made to make the levels look similar to the county names used in the maps package.

population

a numeric vector

top

a character vector

second

a character vector

third

a character vector

Source

U.S. Census Bureau. The ancestry data were extracted from Summary File 3, available from

http://www.census.gov/census2000/sumfile3.html

which is based on the ‘long form’ questionnaire (asked to 1 in 6 households surveyed).

References

http://www.census.gov/prod/cen2000/doc/sf3.pdf

See Also

mapplot, for examples.

[Package latticeExtra version 0.6-29 Index]