postdoc {latticeExtra}	R Documentation

Reasons for Taking First Postdoctoral Appointment

Description

Reasons for Taking First Postdoctoral Appointment, by Field of Doctrate, 1997

Usage

data(postdoc)

Format

The data set is avaliable as a two-way table of counts.

Source

Survey of Doctorate Recipients, 1997

References

Enhancing the Postdoctoral Experience for Scientists and Engineers: A Guide for Postdoctoral Scholars, Advisers, Institutions, Funding Organizations, and Disciplinary Societies

http://books.nap.edu/catalog.php?record_id=9831

Examples

data(postdoc)
library(lattice)
barchart(prop.table(postdoc, margin = 1),
         auto.key = TRUE, xlab = "Proportion")

[Package latticeExtra version 0.6-29 Index]